Move DICOM helper classes to new loci.formats.dicom package
