+++ components/formats-bsd/build.properties
@@ -15,6 +15,6 @@ component.deprecation    = true
 component.resources-bin  = loci/formats/bio-formats-logo.png \
                            loci/formats/utests/2008-09.ome
 component.resources-text = loci/formats/*.properties \
-                           loci/formats/in/*.txt
+                           loci/formats/dicom/*.txt
 
 component.main-class     = loci.formats.gui.ImageViewer
