+++ components/formats-bsd/src/loci/formats/dicom/DicomTile.java
@@ -30,7 +30,7 @@
  * #L%
  */
 
-package loci.formats.in;
+package loci.formats.dicom;
 
 import loci.common.Region;
