+++ components/formats-bsd/src/loci/formats/dicom/dicom-dictionary.txt
